+++ src/org/ods/orchestration/util/Project.groovy
@@ -1923,7 +1923,7 @@ class Project {
                 this.getComponentDiscontinuations(oldData, newData)
             newData.discontinuations = discontinuations
             // Expand some information from old saved data
-            def newDataExpanded = expandPredecessorInformation (oldData, newData, discontinuations)
+            def newDataExpanded = expandPredecessorInformation (oldData, newData)
             newDataExpanded << [discontinuationsPerType: discontinuationsPerType(oldData, discontinuations)]
 
             // Update data from previous version
@@ -2099,7 +2099,7 @@ class Project {
      * @return Map new data with the issue predecessors expanded
      */
     @NonCPS
-    private static Map expandPredecessorInformation(Map savedData, Map newData, List discontinuations) {
+    private static Map expandPredecessorInformation(Map savedData, Map newData) {
         def expandPredecessor = { String issueType, String issueKey, String predecessor ->
             def predecessorIssue = (savedData[issueType] ?: [:])[predecessor]
             if (!predecessorIssue) {
@@ -2126,11 +2126,7 @@ class Project {
                         def expandedPredecessors = predecessors.collect { predecessor ->
                             expandPredecessor(issueType, issueKey, predecessor)
                         }.flatten()
-                        // Get old links from predecessor (just one allowed)
-                        def predecessorIssue = savedData.get(issueType).get(predecessors.first())
-                        def updatedIssue = mergeJiraItemLinks(predecessorIssue, issue, discontinuations)
-
-                        [(issueKey): (updatedIssue + [expandedPredecessors: expandedPredecessors])]
+                        [(issueKey): (issue + [expandedPredecessors: expandedPredecessors])]
                     }
                 }
                 [(issueType): updatedIssues]